% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{launch_tutorial}
\alias{launch_tutorial}
\title{Lance le tutoriel}
\usage{
launch_tutorial(file, port = 8181, host = "localhost")
}
\arguments{
\item{file}{le file Rmd du tutoriel}

\item{port}{le port sur lequel on envoie la page web}

\item{host}{l'url sur laquelle on publie}
}
\description{
Lance le tutoriel
}
